Modularize application code
